requirements.txt
supabase==2.9.1
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn==0.32.1
pydantic==2.10.4
requests==2.31.0 